Feat: Connect SearchDetailFragment and LookPartyFragment
